Move nearby scene feed endpoint to live under `/scene`. Have tessellation name as URL parameter.
